用户controller，问题controller APIPost测试

+++ src/main/java/com/community/mapper/QuestionExtMapper.java
@@ -5,7 +5,7 @@ import org.apache.ibatis.annotations.Mapper;
 
 @Mapper
 public interface QuestionExtMapper {
-    int addView(Integer Id);
-    int addComment(Integer Id);
-    int addLike(Integer Id);
+    int addView(Long Id);
+    int addComment(Long Id);
+    int addLike(Long Id);
 }
